Hook up send_notification signal on Event.post_save
.. and move all signals-handling to the correct place, to negate the cyclical
import potential even more.
