+++ src/argus/notificationprofile/signals.py
@@ -1,0 +1,19 @@
+from argus.auth.models import User
+from .models import TimeRecurrence, Timeslot
+
+__all__ = [
+    "create_default_timeslot",
+]
+
+
+# Create default immediate Timeslot when a user is created
+def create_default_timeslot(sender, instance: User, created, raw, *args, **kwargs):
+    if raw or not created or instance.timeslots.exists():
+        return
+
+    TimeRecurrence.objects.create(
+        timeslot=Timeslot.objects.create(user=instance, name="Immediately"),
+        days=[day for day in TimeRecurrence.Day.values],
+        start=TimeRecurrence.DAY_START,
+        end=TimeRecurrence.DAY_END,
+    )
